pyproject.toml
[project]
name = "advent"
version = "0.1.0"
description = "Solutions for the 2025 Advent of Code challenges."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selectolax>=0.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["advent"]
